# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Trading Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analyses:
    post:
      operationId: createAnalysis
      tags: [analyses]
      summary: Create new analysis (triggers AI)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnalysisBody"
      responses:
        "201":
          description: Analysis created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: AI service error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listAnalyses
      tags: [analyses]
      summary: List user's analyses with filters
      parameters:
        - name: mode
          in: query
          schema:
            type: string
            enum: [beginner, pro]
        - name: instrument
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: from
          in: query
          description: Filter analyses created on or after this date (ISO 8601)
          schema:
            type: string
            format: date
        - name: to
          in: query
          description: Filter analyses created on or before this date (ISO 8601)
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of analyses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysesList"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analyses/summary:
    get:
      operationId: getAnalysesSummary
      tags: [analyses]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysesSummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analyses/recent-instruments:
    get:
      operationId: getRecentInstruments
      tags: [analyses]
      summary: Get 3 most recently analyzed instruments
      responses:
        "200":
          description: Recent instruments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentInstruments"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analyses/quota:
    get:
      operationId: getAnalysisQuota
      tags: [analyses]
      summary: Get current user's analysis quota usage
      responses:
        "200":
          description: Quota usage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisQuota"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analyses/personal-analytics:
    get:
      operationId: getPersonalAnalytics
      tags: [analyses]
      summary: Get personal analytics data
      responses:
        "200":
          description: Personal analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonalAnalytics"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analyses/{id}:
    get:
      operationId: getAnalysis
      tags: [analyses]
      summary: Get single analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /analyses/{id}/feedback:
    post:
      operationId: submitFeedback
      tags: [analyses]
      summary: Submit feedback for analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackBody"
      responses:
        "201":
          description: Feedback submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Analysis:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        instrument:
          type: string
        timeframe:
          type: string
        userInputContext:
          type: string
          nullable: true
        mode:
          type: string
          enum: [beginner, pro]
        validUntil:
          type: string
          format: date-time
        marketCondition:
          type: string
          nullable: true
        riskLevel:
          type: string
          nullable: true
        confidenceMin:
          type: integer
          nullable: true
        confidenceMax:
          type: integer
          nullable: true
        mainScenario:
          type: string
          nullable: true
        alternativeScenario:
          type: string
          nullable: true
        whyReason:
          type: string
          nullable: true
        failureConditions:
          type: string
          nullable: true
        baseCase:
          type: string
          nullable: true
        bullishScenario:
          type: string
          nullable: true
        bearishScenario:
          type: string
          nullable: true
        keyDriversTechnical:
          type: string
          nullable: true
        keyDriversFundamental:
          type: string
          nullable: true
        marketContext:
          type: string
          nullable: true
        invalidationConditions:
          type: string
          nullable: true
        uncertaintyNotes:
          type: string
          nullable: true
        feedback:
          $ref: "#/components/schemas/Feedback"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - instrument
        - timeframe
        - mode
        - validUntil
        - createdAt
    AnalysesList:
      type: object
      properties:
        analyses:
          type: array
          items:
            $ref: "#/components/schemas/Analysis"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - analyses
        - total
        - page
        - limit
    CreateAnalysisBody:
      type: object
      properties:
        instrument:
          type: string
        timeframe:
          type: string
          enum: [1m, 5m, 15m, 1h, 4h, 1D, 1W]
        userInputContext:
          type: string
        mode:
          type: string
          enum: [beginner, pro]
      required:
        - instrument
        - timeframe
        - mode
    AnalysesSummary:
      type: object
      properties:
        totalAnalyses:
          type: integer
        beginnerCount:
          type: integer
        proCount:
          type: integer
        avgConfidenceMin:
          type: number
          nullable: true
        avgConfidenceMax:
          type: number
          nullable: true
        recentAnalyses:
          type: array
          items:
            $ref: "#/components/schemas/Analysis"
      required:
        - totalAnalyses
        - beginnerCount
        - proCount
        - recentAnalyses
    RecentInstruments:
      type: object
      properties:
        instruments:
          type: array
          items:
            type: object
            properties:
              instrument:
                type: string
              lastAnalyzedAt:
                type: string
                format: date-time
              mode:
                type: string
            required:
              - instrument
              - lastAnalyzedAt
              - mode
      required:
        - instruments
    AnalysisQuota:
      type: object
      properties:
        unlimited:
          type: boolean
          description: True for admin/super_admin, who bypass quota
        hourly:
          type: object
          properties:
            limit:
              type: integer
            used:
              type: integer
            remaining:
              type: integer
          required: [limit, used, remaining]
        daily:
          type: object
          properties:
            limit:
              type: integer
            used:
              type: integer
            remaining:
              type: integer
          required: [limit, used, remaining]
      required: [unlimited, hourly, daily]
    PersonalAnalytics:
      type: object
      properties:
        totalAllTime:
          type: integer
        totalThisMonth:
          type: integer
        totalThisWeek:
          type: integer
        topInstruments:
          type: array
          items:
            type: object
            properties:
              instrument:
                type: string
              count:
                type: integer
            required:
              - instrument
              - count
        dominantMode:
          type: string
          nullable: true
        accuracyRate:
          type: number
          nullable: true
        feedbackCount:
          type: integer
        weeklyData:
          type: array
          items:
            type: object
            properties:
              week:
                type: string
              count:
                type: integer
            required:
              - week
              - count
      required:
        - totalAllTime
        - totalThisMonth
        - totalThisWeek
        - topInstruments
        - feedbackCount
        - weeklyData
    Feedback:
      type: object
      properties:
        id:
          type: integer
        analysisId:
          type: integer
        feedbackType:
          type: string
          enum: [useful, not_useful]
        outcome:
          type: string
          enum: [correct, wrong, unknown]
          nullable: true
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - analysisId
        - feedbackType
        - createdAt
    FeedbackBody:
      type: object
      properties:
        feedbackType:
          type: string
          enum: [useful, not_useful]
        outcome:
          type: string
          enum: [correct, wrong, unknown]
          nullable: true
        note:
          type: string
          nullable: true
      required:
        - feedbackType
...